# Elas_Conectadas_BackEnd/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Elas Conectadas",
    "description": "Arquitetura Orientada a Contratos (Modular)",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login efetuado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/auth/send": {
      "post": {
        "operationId": "AuthController_sendMail",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sendEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "E-mail enviado"
          }
        }
      }
    },
    "/auth/request-otp": {
      "post": {
        "operationId": "AuthController_requestOTP",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/requestTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP solicitado"
          }
        }
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOTP",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP verificado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "empreendedora@teste.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "SenhaSegura123!"
          }
        }
      },
      "LoginResponseDto": {
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Token JWT para autenticação"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        }
      },
      "sendEmailDto": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "empreendedora@teste.com",
            "description": "E-mail da destinatária"
          }
        }
      },
      "requestTokenDto": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "empreendedora@teste.com"
          }
        }
      },
      "VerifyOtpDto": {
        "type": "object",
        "required": [
          "email",
          "otp"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "Código OTP recebido por e-mail"
          }
        }
      },
      "UserResponseDto": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "accountStatus"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "dob": {
            "type": "string",
            "nullable": true,
            "description": "Data de nascimento no formato DD/MM/AAAA"
          },
          "occupation": {
            "type": "string",
            "nullable": true
          },
          "rua": {
            "type": "string",
            "nullable": true
          },
          "numero": {
            "type": "string",
            "nullable": true
          },
          "bairro": {
            "type": "string",
            "nullable": true
          },
          "cidade": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "pfp": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "VERIFIED",
              "UNVERIFIED"
            ]
          }
        }
      },
...
    }
  }
}